Decouple sidebar and mailbox by extracting SelectedMailboxLocation
which is reponsible for keeping in memory the currently selected
location as an instance of `MailLocation` (which is supposed to be only
used to represent the currently selected location _and should not be
used for a message's location instead_).

Also introduce SidebarViewModel which has the responsibility to create
and update the state of the sidebar.

Extact uiModels to their own files and packages

MAILANDR-67