Update core libraries to include new MailSettingsRepository implementation
The repo was modified to allow settings to be changed while offline and
work optimistically. MailSettings DI was also moved to a own dagger
module, so explicit MailModule is not needed anymore.

This solves all the issues mentioned in the previous commit and
re-enables the acceptance test introduced with `90406ceddc46efa`
to mark the functionality as complete.

MAILANDR-111