Refactor MailboxViewModel state to be internally mutable
The purpose of this change is to have the ability to change the state
when actions are performed on the UI (eg. an item is selected / the
unread filter is enabled....) without the need to have an internal
MutableFlow for each of those actions.

This forces us to use the `init` to start observing data. The way state can
change are two only, which is through a change in any _data_ (which will
cause a new emission from the flow in the `init`) or thorugh an user
action which is received in the `submit` method and "reduced" to a new
state.

MAILANDR-134