When location changes scroll to mailbox top
This feature was initially introduced with MAILANDR-149 and then broken
with the MailboxVM state's refactoring done by MAILANDR-134.
This solution uses the concept of effects that we meanwhile introduced,
in the attempt of only applying this "scroll to top" animation in the
cases where the label did actually change.

**Note that one test is purposefully failing in this commit:** this is
pointing out one case where the current approach is not working in order
to gather feedback on the PR.

MAILANDR-134