Use androidx orchestrator to clear app data between tests
This helps isolating tests and making them more stable.
It was introduced to solve a crash that appeared sometimes in UI tests
after updating Core libs from version 8.1.0 to 8.4.0 and was making them
flacky. While this is making the tests work consistently, the root cause
that introduced the crash is still being investigated as it might not be
fully fixed by this change alone. More details in the task.

MAILANDR-182