Remove problematic SmokeTest annotation
This annotation was meant to declare tests to run as part of "SmokeTest"
suite, but it isn't working well.
It looks for new classes to be considered by SmokeTest suite, they need
to _both_ be declared as part of  `SmokeSuite@SuiteClasses` annotation
and contain tests that are annotated as @Category(SmokeTest).
This is redundant and confusing, thus we drop the SmokeTest.