Location icons are only shown on mailbox items when in certain locations
which are Starred, All Mail, Custom Labels and search results. The
latter was not taken into consideration in this changes as not yet
implemented.

Here the use case performs the check on the current location and the
view model maps the result to a list of icons (res ids) to be easily
displayed in the view

MAILANDR-70