Workaround strict mode whitelisting issues on API 31
The method to `getStackTrace` used for the whitelisting "hack" is not
available on all API levels (ie. doesn't seem to exist on API 25, is
annotate as @TestApi on API 31).
When we are not able to `getStackTrace` to check whether the reported
violation is in the whitelist, we go on and report it anyways. This
requires to log it as a penalty instead of crashing (which would make UI
tests unreliable depending on device)

MAILANDR-135