Introduces a reducer for mailbox top app bar state
On top of adding the reducer, also introduces an operation interface
which has two subtypes: viewaction, triggered as an action coming from
the view, i.e. user interaction with the app, and an event, which is a
result of emissions that the view model observes.

Each of the operations can implement multiple Affects* interfaces. This
way, each sub-reducer needs to only worry about operations that affect
it (and have an exhaustive when without an else, that only checks the
operations relevant to this part of the mailbox state).

MAILANDR-272