A conversation containing a trashed / spam draft is considered as trashed
This is a followup of commit `7f40f32bbeddae2b`. When checking whether
all messages in a conversation are trash or spam, we look at the
conversation's labels: in case a draft was moved to trash / spam, we
would have "all-drafts" label in the conversations label ids list.

The presence of "all-drafts" (as well as "all-mail" / "all-sent") labels
does not make a difference for the goal of checking whether all the
messages are trashed / marked as spam, thus we ignore them.

MAILANDR-221