Fix Flow tests failing due to splashscreen blocking UI  thread
When updating to core version 9.8.0 (PR#263) the "flow" tests started
failing with the following error:
 > Could not launch intent Intent { act=android.intent.action.MAIN cat=[android.intent.category.LAUNCHER] flg=0x14008000 cmp=ch.protonmail.android.dev/ch.protonmail.android.MainActivity } within 45000 milliseconds. Perhaps the main thread has not gone idle within a reasonable amount of time?

which resulted being caused by the splashscreen
`setKeepOnScreenCondition` blocking the UI thread due to being
continuously evaluated. This somehow prevented the `Launcher` composable
from retrieving any state updates from its ViewModel, which in turn
caused the redirect to the "login" activity not to happen.

While the root cause for this to start failing was not pinpointed
with confidence, the main suspect is the update / addition of some compose
dependencies (compose 1.2.0-rc02 -> 1.3.2, compose-foundation 1.3.1 and
compose-material 1.3.1 added).

The fix consists in changing the behavior of the splashscreen to only
stay visible while the `LauncherViewModel` is being initialised
(represented by the `Processing` state) and remove it afterwards.

This seems to be a better behavior than the previous one, as instead
of leveraging the fact that launching the Login activity would pause
`MainActivity` (effectively freeing the UIThread from the splashScreen
stalking) it actually only keeps the splash screen shown for the minimun
needed time and then delegates to the "Launcher" the responsibility of
choosing what to show.

MAILANDR-360