Already shown conversation metadata is not removed when errors occours
This avoids that the metadata of the conversation in the
detail screen (eg. subject, num of messages and starred state) gets
hidden when receiving an error from the remote call to refresh / fetch
messages.
Such error will be handled by the messages reducer instead and shown in
place of the messages.

MAILANDR-358