Handle API failures represented with a nested "ProtonErrorException"
In some cases API calls failures are returned as a
"ProtonErrorException" nested inside an APIException inside the data
result.
ProtonErrorException is used to represent multiple things, spanning from
404 http errors to domain specific errors (eg. Attachments upload state).

This commit puts in place handling for some common cases and avoids
crashing when an unknonw case is received: instead, it will be returned
as "unknonw" and logged to Sentry.

MAILANDR-265