Introduce MockPriority
To better handle cases where exact matches need to be performed while wildcard matched requests are still defined, a priority has been introduced for mocks.

Example:
`/mail/v4/conversations/*` needs to be wildcard matched to load a conversation detail screen, but at the same time ``/mail/v4/conversations/read` needs to be called and handled with a different payload. If we define a higher priority for the latter, since we don't know how many calls are required with wildcard matching, we prioritize the exact match above all and whenever the /read call is performed, a different response compared to the wildcard match one is served.