Conversation in-memory repository
Added a new InMemoryConversationStateRepository to hold a cache of the messages state in a conversation (collapsed, expanding, expanded) and related use cases that can be used from a VM to keep track of the message state of all the messages in a conversation. The repository is scoped to the view model with hilt so that every use case that uses it in a view model instance will receive the same repository instance.

The expanded state will keep a reference to the decrypted message bodo so that decryption is done only once per message while the conversation screen is open.

This code will be later used from the ConversationDetailViewModel.

MAILANDR-545