Extract methods to make pagination's prev / next key logic more explicit
The key logic is complicated to understand due to how pagination works
underneath: the paging source is expected to return a null key (
either previous or next) when we do not have data locally to satisfy the
request. When a null key is returned, then paging-library internals will
cause the remote mediator to be triggered for the key that was
originally requested, which will eventually fetch such data.

The logic in place, which is now abstracted by the
`doNotTakeWhenMediatorShouldBeTriggeredTo..` methods is there to avoid
triggering remote mediator when not needed (eg. we never featch items
to "prepend" when the current operation is an append) but at the same
time trigger it when it's needed)

MAILANDR-410