Update FormatShortTimeTest
Due to changes introduced with CLDR 38 support, "September" is shortened as "Sept" in `Locale.UK` rather than "Sep", causing a test to fail due to the expected value not being the actual.

The itemTime base value has been changed to use October instead, as it does not introduce inconsistencies between JDK versions.

MAILANDR-565