Lift showing offline / error snackbars to MailboxSwipeRefresh composable
This is more correct than triggering the snackbar as part of the
MailboxItemsList composable `when` condition, which is used within an
`item` to actually show composable elements related to the "append"
state (eg. loading, error).

It also fixes a case where the snackbar (both offline or error) would
not show even if expected to at pull-to-refresh with existing data
already shown, due to the LazyColumn scope not being recomposed (as list
state didn't change).

MAILANDR-561