Let mailbox refresh logic be handled by the pagination lib alone
By dropping any calls to the ViewModel and any custom logic that was
performed.
The trigger for this change was the "offline" snackbar being
shown twice when performing a pull to refresh, once immediately after
pulling (triggered by the `viewModel.onRefresh` method being called and
causing an emission of the `NetworkStatusRefreshed` which in turn
displayed the snackbar (both the offline and the error ones, as this
emission also caused recomposition of the MailboxSwipeRefresh
composable)) and once when the actual result of the refresh was returned
(from paging library through mediator / pagingSource).

The concept of NetworkStatusRefreshed effect was introduced to ensure
the offline snackbar was indeed shown when performing a pull to refresh:
given that's now handled by the pagination lib, none of this logic is
needed anymore.

MAILANDR-561