Do not show offline / error snackbars when opening a detail
Due to recomposition of MailboxSwipeRefresh composable (triggered by the
emission of the openItemEffect used to show the detail itself), the
offline / error snackbars are shown twice each time a detail screen is
opened.

LaunchedEffect is a workaround that prevents this from showing when
opening the detail, even though it will still be shown when coming back
to the mailbox.

MAILANDR-561