Avoid crash when the very first "get labels" request fails
This can be reproduced by logging into the app and enabling flight mode
right away. The `getLabels` request which implicitly happens when
loading the mailbox page will fail (no network) and throw an exception,
which is not handled anywhere and causes the app to crash.
More detailed steps to reproduce in the task.

This commit handles this exception as a generic mailbox error, meaning
no data will be shown even if available locally (which might be the case
if the /getMessages OR /getConversations requests that happen in
parallel through the mediator succeed while the /getLabels fail).
This is assumed to be an edge case and thus not considered yet. If later
on we want to allow showing existing data even though labels fail, this
solution allows to easily introduce such change.

MAILANDR-561