Flatten Composer draft state for simplicity
The tracking of the Submittable and NotSubmittable state was flattened
to be a boolean flag of a `ComposerDraftState` data class which holds
all of the data that the composer needs.
This simplifies the work of the reducer to handle the different states,
as there's no need to check for what state we're currently in before
changing any bits of it.

Removed the Submittable / Not submittable wrappers over ComposerForm
and used the form directly (as after the previous change that lifted
the error they had no differences left).

MAILANDR-487