Move get composer sender address use case call from reducer to VM
Putting the use case in the reducer was an attempt to try avoiding some
downsides of having it in the VM (eg. having to create additional events
for each possible outcome of the use case, reducer handling actions that
do nothing such as OnChangeSender..) but eventually we found out that
there are more downsides in exposing the reducer to "side effects".
(consistency with the other reducers was also a consideration).

MAILANDR-487