Remove separated user-error for failure to resolve user address
Same as previous commit, resolving a user-selected email address is
never expected to fail (given the constraints ComposerScreen provides to
allow selecting one). In the case of failure, no actionable feedback to
provide to the user is currently known.
For the sake of simplicity, we then reuse the existing save draft error
to provide feedback.

MAILANDR-603