Create API client for device registration
When we get a firebase token, we need to send it to the API to associate it with the user ids of the accounts logged in in the device.

Here we add the API definition, the worker that calls the API and the DataStore to store the token for later use.

MAILANDR-751