Extract logic from repository to a use case to sync draft
This allows makings space for a worker to be introduced between the
repository and the use case.

Changed return type of syncDraft use case and repository to be either
DataError or Unit. This is telling clients not to await the success
object directly from the use case (as it's run async and might happen at
any time in the future once the worker is introduced) but to observe DB
instead.

MAILANDR-491