Add a deep links handler view model to help with managing the deep links from compose. We can not add the deep links directly to the different routes because we are using the `savedStateHandle: SavedStateHandle` in the view models and that is injected by hilt. And the `hiltViewModel()` extension does not allow us to override the extras (the `viewModel()` extension does).
As we can not use the hilt extension directly, we use an intermediate view model that will register all the deep links urls and pass it the values when we receive a deep link for it to decide show to proceed (in the HomeRoutes).

A response to the VM state is just a regular navigation with the right parameters so further VMs can use the `saveStateHandle` as usual.