Move logic to observe update of the message id to apiMessageId away from VM
This logic does not belong on the UI, which can instead keep referencing
the message with the messageId it got when the composer started.
If the message is created with the API in the meanwhile and its id is
updated in the DB, the domain layer will take care of coordinating the
resolution of which is the new message ID needed to perform any network
operations based on the mapping which exists in DraftState.

MAILANDR-491