Fix calculations for labels that need to be removed/added in RelabelConversation use case
In the scenario where a label is partially selected and then gets fully selected,
we were doing more operations than necessary. First, we were removing the label completely,
before adding it again. With the changes in this commit, in this scenario,
we're only adding the label that was fully selected after being partially selected.

MAILANDR-836