Fix crash when updating drafts mailbox right after updating a draft
The crashed (detailed in the below task) was happening when having a
single draft, opening it, closing composer and performing a pull to
refresh (or awaiting for the event of the message update).

The root cause is the time of the message changing, which is a direct
consequence of the update. Given pagination fully relies on time of the
messages, the update was causing:
- When update happens in DB, pagination is triggered as the interval is
  invalidated
- RefreshKey to generate a Key specific to the time of the only message
  in the list (which is not outdated)
- getting messages from DB for such refresh key results in 0 hits,
  as the key is contrained to a single timestamp which doesn't exist
  anymore in DB
- Finding no messages with the refresh key causes both next and previous
  keys (generated by getAdjacentKeys use case) to be generated without
  any time bounds (-INF -> +INF) as there is no item in the list to be
  used ad bound (which is why this doesn't happen if there are more than
  one items, as the next one in the list will work as bound)
- Both next and prev keys will find the same item (which is the draft
  that got updated) and return it as part of different pages
- Merging the pages and showing them in the view will crash due to the
  same item ID being there more than once

The fix, consists in detecting when a refresh key is being generated for
a single item and increasing the maxTime value allowed: this will result
in the refresh key working as if it was a "prepend key", looking for
messages from the time of the draft up to +INF (thus, including the
updated draft)

MAILANDR-854