Avoid launching test subject in a separated coroutine (POC)
Move the "when" condition of tests using turbine outside of turbine
"test" method, to avoid exceptions that are thrown in the scope to fail
the test.
This was needed due to last core update transitiently updating
coroutines, which will now fail when an exception is thrown during a
tests execution in any scope that's not the test one.

CP-6207