Send message work chain uses the same unique name as draft upload
This solves a possible (and likely) race condition between the "standalone"
draft upload (which is scheduled to happen every second, so with most
probablilities will be ongoing while the user hits "send") and the
upload of the draft that happens of part of the sending.

Using the same "uniqueName" (or workerId) leverages the workers to
ensure this operation only happens once at a time.

MAILANDR-657