Rely on vm internal logic to decide when to change draft state
When HomeViewModel receives draft states telling that a message was
send (or failed sending), we show the feedback to the user and update
the "draft state" in the DB to ensure the same feedback isn't shown
again (for a sent message we delete draft state while for an errored one
we reset it back to "Syncronized").
This was done in a first approach by letting the UI send an "action" to
the VM after it did show the fedeback, which was a workaround to avoid
the emission caused by updating the draft state to prevent the feedback
from being shown.

A better solution added in this commit, is to handle this in the VM by
mimicing the reducer logic (ie. checking the current state and avoiding
resetting the effect if it's currently set). As a shortcut, we avoid
directly to change the state when the received drafts state do not
require showing any feedback.

MAILANDR-657