Fix attachment observing while composing message
Fixes the issue with disappearing attachments. This was caused by the changed message id when a local draft got synced and the local message id got updated by the id received from remote. To avoid this, the draft state is now observed which will adapt which id is used for observing the connected attachments

MAILANDR-661