Fix network logging in release builds
Because of the enableDebugLogging flag set to true permanently, logging is also happening in release builds. This caused issues when uploading large files on low-end devices, because of the logging, the whole file is kept in the buffer. Without logging, which is in release builds not needed anyway uploading of large files works without an issue on low-end devices.

The issue to that can be found here: https://github.com/square/retrofit/issues/540