Do not drop blank lines when reply / fw to a message
Any new lines introduced by the user were lost due to the message's
mimetype being changed to "Html" when a "quote" is added to it (which
is, when replying or forwarding a message).
Wrapping the user's typed-in text in a `<pre>` tag solves this as
new lines will be preserved when rendering the html.

MAILANDR-1024