Focus draft body field when replying / forwarding an html draft
In this case, we do not have any text prefilled into the body (which was
the previous condition to pre-select it and happens when reply / fw
plain text messages) but we do have a "quoted html" field.
We use the existence of this "quote" field as second condition to trigger the
focusing of the body field.

MAILANDR-1002