Fail gracefully when use case to obtain send preferences throws
This case was seen in the context of MAILANDR-976, where an exception
was being thrown from within the obtainSendPreferences use case (due to
missing minification rules for ez-vcard lib).
While that issue was solved, it is still the case that if an exception
is throw here, sending process would fail "silently" and the user
wouldn't know.
This commit add handling (generic catching as the use case is an
external dependency and we don't know what it can throw) and returns an error
that will cause feedback to be shown to the user.

MAILANDR-979