Add ForegroundAwareAutoLockHandler and related initializer
The handler is used to pick up background state changes and update the auto lock last foreground millis to the current Instant.

The collection of the background state is delayed to prevent the initial background state to be picked up when opening the app from scratch (as it goes from Background -> Foreground when the app is not in recents).

MAILANDR-1341