Allow loading remote content for a specific message body
When the user's preference is to block remote content from being loaded
when displaying messages, we show a banner at the top of the message
body to notify them the content was blocked.
Now, we also show a button that allows loading it for this specific
message.

The webview needs to be fully recomposed in order for the change to take
effect, as without re-creation it wouldn't pickup the new
"webViewClient" which is the one that takes the decision on whether to
block the remote content.

MAILANDR-1074