Allow loading embedded images for a specific message body
When the user's preference is to block embedded images from being loaded
when displaying messages, we show a banner at the top of the message
body to notify them they were blocked.
Now, we also show a button that allows loading them for this specific
message.

The webview needs to be fully recomposed in order for the change to take
effect, as without re-creation it wouldn't pickup the new
"webViewClient" which is the one that takes the decision on whether to
block the embedded images.

MAILANDR-1075