Solve crash due to pagination keys returning the same item twice
This issue is a flavour of MAILANDR-854,
the root cause still being that the pagination library performs a request
using the same MailboxPageKey as both Previous and Next keys,
which in turns return the same item on both pages crashing when trying
to display them in the lazy list.

The tentative solution consists in excluding the previous key from the
Page when it's exactly the same as the next one. This has the potential
of creating an unneeded network call in some cases (as it would trigger
the mediator) in exchange of avoiding the crash.

MAILANDR-1419