Fail gracefully when draft upload fails due to message already sent
We might get in this state due to a race condition (in between clients
or due to concurrency issues locally).
When the API calls fails with such error code (422http, 15034proton)
then do not notify this to the the user as a sending failure but set the
draft state as sent and interrupt the sending chain of work.

MAILANDR-1223