Save api-defined conversaton id to local message when draft is created
When a new draft is created locally, it does not have a conversation id as
it's a backend-defined information.
The lack of such bit of data causes a crash when:
- conversation mode is enabled
- users composes and sends a new message
- Device is online and sending operation starts ("Message sending" is seen)
- Before sending completes ("Message sent" seen) the user navigates to
  the "sent" folder and opens the message
(some alternative steps lead to the same crash too)

The issue happens as we try to navigate to a conversation, but there's
no conversationId to go to. It's fixed automatically once the draft
creation (First step of the sending process) completes and the assigned
conversationId gets back to the app (through event loop or user refresh).

This fix stores the conversationId locally beforehand, the first time
the draft is created (which with device online happens right away when
the user starts typing a draft's body).

MAILANDR-982