Composer form is inside a Column with reverse scrolling. This means in order to scroll to the top we need to know the size of the whole composer form and scroll to its height.
This commit provides a general solution without calculating the ComposerForm height or handling WebView loading etc.

We use Column's scrollState.maxValue to scroll to the top of the composer screen (until user starts editing the body). When the WebView is loaded with the quoted html this value also gets updated.
In addition, this commit also provides scrolling to top solution when there is no quoted text like opening a draft message for example.

MAILANDR-1492