LazyColumn data is provided by LazyPagingItems and list of items can be updated from local cache first and then from remote source. It does not scroll automatically to show newly added items to the list.
Proposed solution: With this commit we monitor the listState.firstVisibleItemIndex until the user scrolls the mailbox. ListState may frequently change therefore we only monitor it until the user taps the screen and scrolls the mailbox. If a new item is received, listState.firstVisibleItemIndex gets updated and then we scroll to the top of the mailbox,

MAILANDR-916