Avoid composer to scroll to bottom when responding inline
Due to the composer's usage of reverseScrolling, when we inject large
bodies in the composer text field (as part of the "respond inline"
feature converting large html quote to plain text) the composable
scrolls to the bottom of the inserted text (as if the user would be
typing it in).

This hack scrolls back to the top of the view after the body was
injected, shadowing the approach to keep scroll at the top when loading
the quoted content intially.
Some better alternatives such as setting the "reverseScrolling" property
dynamically (making it false while injecting and then true again) didn't
work out as setting it to true again still causes scroll to bottom to
happen.

MAILANDR-1699