Update basic project automation for the ET project fork
- Disable CI stages that shouldn't run on ET project
- Point firebase deploy scripts to a newly created "et" app
- Update base package name to "ch.protonmail.android.et" to allow
  running in parallel with v6

Context: ET (aka "enigineering transformation") project is a PoC where
the domain and data layers are provided by a shared libarary written in
Rust.
This repo is a fork of the production one (android v6) done to apply
such concept on a higher-fidelity manner than the quick-and-dirty PoCs
that were done so far.