Replace mailbox data module with data-rust
This is the first step in the transition from the local data handling
(Room, Retrofit, ProtonCore libs) to the core rust library which will be
shared cross platforms.

Striving for this transition to be as step-by-step as possibile,
the tentative way to tackle it is:

- Duplicate mailbox-data module to data-rust;
- From such new module, drop all deps related to room, core and networking;
- Add the rust lib dependency to the new module (_not done by this
  commit as not applying to mailbox module_);
- Add implementations for the Repositories interfaces (living in domain,
  target leaving them untouched) which will fetch data from the rust lib;
- Instruct the build system to ignore the old data module (remove from
  settings.gradle);

While ideally the above steps are applicable one module at a time, in
reality cascading changes are often needed due to infra modules
dependencies. In that case:
- Duplicate other data-modules in cascade
  - eg. common-data --> common:data-rust;
  - From such new module, drop all deps related to room, core and networking;
  - [In case it's not feasible yet to disable the old data module]
    Remove everything that's not yet needed from such modules
    (avoids duplicated classes failures);
  - Add the (common:)data-rust dependency to other modules which are
    relying on (common:)data;

ET-308