Introduce mail-message-rust module and MessageRepository Rust implementation
Following the approach outlined in the previous commit, this one adds a
data-rust module which provides a (fake) implementation of
MessageRepository.
This is the place where rust library can be introduced, any adaptation
in the data types should happen within this module allowing external
clients of MessageRepository to be unchanged.

The dagger bindings where changed to provide this repository, which
means the app will now fail where previously messages were being loaded
- this can be changed simply by switching back the binding to the
previous one.

The other bindings (local / remote data sources are left in place due to
the unfortunate dependency on them by other modules: this should allow
the migration to be done progressively, with the app building and
running tests at all times. Eventually, when all usages are switched to
data-rust modules, the whole data one should be disabled and the
bindings dropped.

ET-308