Make injection of Rust data layer depend on a build time flag
This allows to build the project using the "local" Room data layer or
switching to the rust provided one.
The latter, is under development and (at this commit) implmentations are
mocked (leading to crashes at runtime). Making this dynamic allows to
keep both the old and new modules enabled (to import new changes from v6
without braking) and to keep UI tests on CI running.

ET-308