Create RustSessionRepository to hold rust library static session
This "session" object ("MailSession" in the rust lib) is the first
object that clients has to initialize in order for the library to work.
It exposes a mix of user and app management methods (connectivity,
login, getCurrentUser, flush quque etc).

To try creating a better abstraction, this repo would receive the
"mailSession" generated from the lib initializer (issue#1: through a setter,
which is quite fragile) and hold it in memory (static injector) to allow
it to be reused by other repos (eg. UserRepository can be created for
the sake of good abstraction and will need session to operate).

issue#2: This abstraction forces us to expose rust models in the domain
layer (through the Repositories interfaces)

ET-787