User Session Repository handles callbacks from Rust to update session
In this basic handling, when we receive errors or failures from rust we
"revoke" the user's session by emitting null on the flow. This should
cause the user to be redirected to the login page and should instead be
ignored by the other flows that depend on the session (eg. observing
messages etc)

ET-787