Abstract core lib LabelRepository behind an app-controlled interface
This allows to re-use the same approach as for message and conversation
repositories and inject either the "local" (core) version of it or the
"rust" one.
The interface was used in many places, causing a bulk of import changes.
By default, the new CoreLibsLabelRepository is injected and it works as
a middleware to core's LabelRepository (effectively, keeping the
behavior equivalent to before the change).

ET-799