Initialize user context from login flow after first login
This operation is essential for the rust lib to get correctly
initialized. Missing it or performing operations on MailSession objects
before this init is comlete will lead to broken behaviors very difficult
to diagnose (empty values, no errors, no logs).

This should be moved to the login flow once implemented.

ET-804