Implement mail settings repository to read data from rust lib
Only the observe and get methods are implemented. The additional ones to
edit mail settings will throw exception as they are not expected to be called
by production code (settings will be exposed in a webview).
We're forced to keep them in place due to reusing core's interface, they
can be dropped once all the modules are switched to using the rust
implementation and we drop CoreMailSettingsRepository.

ET-441