Use case to observe primary user id can observe rust user sessions
Through allow observing current userId on UserSessionRepository.

This allows us to keep in place all the usages of observePrimaryUserId
at the flow's "initialization point" (ie. VMs) to observe the current
user ID and "switch" the flows when a new one is emitted.

An implementation to allow using the old core accountManager was added
for compatibility when building with the "rust flag" off.

ET-829