User session allows getting a session by userId
This will be used downstream from the `observeCurrentUserId` use case to
get the session for the currently active user.

observeCurrentUserSession method was left in place to allow for current
usages to compile so that they can be replaced in smaller steps by the
following commits. It will be removed shortly.

ET-829