Remove deprecated hardcoded system labels from MailLabels
and adapt all the functionalities that depended on those to depend on
dynamic system labels instead.

As part of this change, some bits of logic are broken and will need
adaptation of whole re-implementation, which will be done in subsequent
commits for the sake of separating changes.

Here's a list of such bits, some of which will need consideration and
some that are just good to know (for logic that is planned be taken care
by rust we don't need to fix it but can just drop it when getting there)

- SelectedMailLabelId will need to pick dynamically which is the first selected label
- when current label is deleted, navigate back to inbox needs adaptation
  (we don't directly know what's inbox label id) (MailboxVM)
- In GetMailboxItemLocation we loose the "should show icon" logic
as we can't check current label agains static system ones to define location
- when HomeVM navigates to drafts, we need to observe mail labels to know the id
(otherwise, we'll re-navigate to drafts even if already in drafts)
- Hardcoded allmail label in ConversationsMapper.getDefaultLabel() method
(not sure if an issue as such fallback shouldn't be needed anyways and
might go away with next rust lib changes)

ET-362