Avoid crash when opening app when there's no "core" session
With b43973fc19 we can now log into the app with a rust-session only.
This causes some components relying on the previous core-based data
layer to throw exceptions and crash the app - in this instance,
PaymentManager used by the sidebar to define whether to show the
subscription button (isSubscriptionAvailable(userId)) crashes when
there's no user.

It seemed overkill to create the same abstraction as we did for other
models to allow reading such data from both rust and core as this bit
will be owned by account team and (as the current assumption)
implemented as a shared-by-platform additional module.

Disabling showing subscriptions till we get such implementation to avoid
waste.

ET-362