Add annotation to allow filtering tests whose logic was implemented in rust
At this commit, no test is included in this annotation (RustReadyTest)
as all SmokeTests crash at launch when trying to perform a login to
rust lib using black env credentials.

Once that's sorted, the plan is to add only tests for which
functionality is already imolemented using rust lib's provided logic in
order to keep the pipeline real and green.

ET-982