Allow fake rust login to retry when failing
This is needed for UI tests, which through MainInitializer do run the
FakeRustSessionInitializer before the mock webserver is even setup which
would cause a lock.
The hack of [making this async +] allow it to fail and retry makes the
trick as the test setup can be completed in parallel.

_edit: the "making this async" part was rolled back as it negatively
impact running the app normally and it doesn't fully solve the UI tests
issues anyways. They will be tackled separately from these changes_

ET-982