Make fake rust login async and allow it to retry
This is needed for UI tests, which through MainInitializer do run the
FakeRustSessionInitializer before the mock webserver is even setup which
would cause a lock.
The hack of making this async + allow it to fail and retry makes the
trick as the test setup can be completed in parallel.

ET-982