Mailbox reduces open item requests to alwyas open conversation detail
This is a decision taken in agreement with product which targets fully
dropping the MessageDetail with the gain of reducing effort (only
one detail screen).
The behavior for the user will be the same as it's now when opening a
message from a message-only location (eg. sent) while in conversation
mode, while it will be a slight difference for users who had
message-only mode.

ET-1024