Rename ItemDetailsOpened event and remove uneeded viewMode parameter
Given now we always open conversation details independently of viewMode,
we don't need the param and can flatten the reducers logic to a single
case (always open conversation using the item's `conversationId`).

The case of `isSentMessageWithNoAssignedConvId` opening message detail
can also be deleted, as we now operate with rust-generate local ids,
thus the case of a newly created message not having a conversation ID
won't be valid anymore.

ET-1024

Changes to MailboxListReducer and test due to rebasing on the addition
of "filter by type" param