Adapt label data source to new API from rust lib
Note that system labels functionality is broken as of this commit due to
the removal of "remoteId" field that was used to define which
SystemLabelId corresponds to which local label Id (which enables
localization and icon mapping in the sidebar).
An enum to fix this functionality will be exposed as a field of rust's
`label` in a future version.

Fixed imports and fields names in UnreadCountersRepositoryImpl
RustConversationMessageQuery

ET-1031